add X-VLA

+++ src/lerobot/scripts/lerobot_eval.py
@@ -191,7 +191,7 @@ def rollout(
         action = action_transition[ACTION]
 
         # Convert to CPU / numpy.
-        action_numpy: np.ndarray = action.to("cpu").numpy()
+        action_numpy: np.ndarray = action.to("cpu").float().numpy()
         assert action_numpy.ndim == 2, "Action dimensions should be (batch, action_dim)"
 
         # Apply the next action.
